front: carve semantic types into front::semantic module
Move `SemanticProgram`, `SemanticFunction`, `StructFieldRegistry`,
`AstGraphOptions`, `FlowingError`, `LoweringAbort`, `ProgramMetadata`,
`qualify_type_name_with_imports`, and `is_path_suffix` from
`front/ast.rs` into a new `front/semantic.rs`.  The types do not
depend on the syn-tree walker; carving them out is the first step
toward retiring the AST graph builder (issue #97 Steps 6.E/6.F).

`front::ast` retains `pub use` re-exports for the moved items so
existing callers compile unchanged; `front::mir`, `jit_codewriter::
call`, and `jit_codewriter::policy` are updated to reference
`front::semantic::*` directly.  No behaviour change; 2786 lib tests
pass on both `mir-frontend` default and `--no-default-features`.
