+++ majit/majit-translate/src/front/mod.rs
@@ -111,9 +111,10 @@
 pub mod ast;
 pub mod mir;
 pub mod raise;
+pub mod semantic;
 
 pub use ast::{
-    AstGraphOptions, SemanticFunction, SemanticProgram, StructFieldRegistry,
     build_semantic_program, build_semantic_program_from_parsed_files,
     build_semantic_program_from_parsed_files_with_statics,
 };
+pub use semantic::{AstGraphOptions, SemanticFunction, SemanticProgram, StructFieldRegistry};
